requirements.txt
requests
fastapi
uvicorn[standard]
slowapi
pillow
python-dotenv
gunicorn
debugpy
redis[hiredis]
flower
celery
httpx[http2]
aiohttp[speedups]
aioodbc
azure-storage-blob